# backend/openapi/auth.yaml
openapi: 3.1.0
info:
  title: LocalTalent Auth API
  version: 0.1.0
  description: >
    Prompt 2 minimal local-account JWT authentication API. This file only covers
    identity login for candidate/company/operator. It does not define RBAC,
    data-scope, field-level permission, consent, withdraw, publish snapshot, or
    integration business flows.
servers:
  - url: http://localhost:8080
paths:
  /api/auth/register:
    post:
      summary: Register candidate or company account
      description: Operator accounts are internal only and cannot self-register.
      parameters:
        - $ref: '#/components/parameters/TraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration succeeded.
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponseEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/auth/login:
    post:
      summary: Login with local account and password
      parameters:
        - $ref: '#/components/parameters/TraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login succeeded and returns a Bearer token.
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/me:
    get:
      summary: Read current authenticated identity
      description: Returns only the minimal identity summary. No permission claims or raw candidate profile data are returned.
      parameters:
        - $ref: '#/components/parameters/TraceId'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current identity.
          headers:
            X-Trace-Id:
              $ref: '#/components/headers/TraceId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityResponseEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TraceId:
      name: X-Trace-Id
      in: header
      required: false
      schema:
        type: string
  headers:
    TraceId:
      description: Echoed or generated trace id for request correlation.
      schema:
        type: string
  responses:
    BadRequest:
      description: Validation error.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Missing, invalid, expired, or tampered credential.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Account already exists.
      headers:
        X-Trace-Id:
          $ref: '#/components/headers/TraceId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    RegisterRequest:
      type: object
      required:
        - identity_type
        - password
      properties:
        identity_type:
          type: string
          enum: [candidate, company]
        mobile:
          type: string
        email:
          type: string
        password:
          type: string
          minLength: 8
          maxLength: 128
          writeOnly: true
        display_name:
          type: string
          description: Candidate display name.
        company_name:
          type: string
        license_no:
          type: string
        user_name:
          type: string
          description: Company user display name.
    LoginRequest:
      type: object
      required:
        - identity_type
        - account
        - password
      properties:
        identity_type:
          type: string
          enum: [candidate, company, operator]
        account:
          type: string
          description: Candidate/company mobile or email, or operator username.
        password:
          type: string
          writeOnly: true
    Identity:
      type: object
      required:
        - identity_type
        - user_id
        - display_name
        - status
      properties:
        identity_type:
          type: string
          enum: [candidate, company, operator]
        user_id:
          type: integer
          format: int64
        company_id:
          type: integer
          format: int64
        display_name:
          type: string
        status:
          type: integer
    LoginData:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
        - identity
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
        identity:
          $ref: '#/components/schemas/Identity'
    RegisterData:
      type: object
      required:
        - identity
      properties:
        identity:
          $ref: '#/components/schemas/Identity'
    RegisterResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RegisterData'
    LoginResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/LoginData'
    IdentityResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Identity'
    SuccessEnvelope:
      type: object
      required:
        - code
        - message
        - trace_id
        - data
      properties:
        code:
          type: string
          example: '0'
        message:
          type: string
          example: success
        trace_id:
          type: string
        data:
          type: object
    ErrorEnvelope:
      type: object
      required:
        - code
        - message
        - trace_id
      properties:
        code:
          type: string
          enum: [VALID_400, AUTH_401, AUTH_409, SYS_500]
        message:
          type: string
        trace_id:
          type: string
        data:
          type: object
          nullable: true